Remove playwright config complexity
RISDEV-6806
