+++ frontend/package.json
@@ -15,9 +15,6 @@
     "test": "TZ=UTC vitest run",
     "test:unit:watch": "TZ=UTC vitest watch",
     "test:e2e": "playwright test test/e2e",
-    "test:e2e:caselaw": "playwright test test/e2e/caselaw",
-    "test:e2e:norms": "playwright test test/e2e/norms",
-    "test:e2e:shared": "playwright test test/e2e/shared",
     "test:a11y": "playwright test --project=a11y",
     "test:queries": "playwright test --project=queries --reporter=./test/queries/queries_log_reporter.ts",
     "coverage": "TZ=UTC vitest run --coverage",
