+++ frontend/test/e2e/caselaw/authentication.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from "@playwright/test"
-import { caselawTest as test } from "../caselaw/fixtures"
+import { caselawTest as test } from "./fixtures"
 
 test.describe("authentication", () => {
   test("name and documentation center should be displayed", async ({
