+++ frontend/test/e2e/caselaw/documents/file-upload.spec.ts
@@ -1,8 +1,11 @@
 import fs from "fs"
 import { expect } from "@playwright/test"
-import { navigateToAttachments, uploadTestfile } from "../e2e-utils"
+import {
+  navigateToAttachments,
+  uploadTestfile,
+  createDataTransfer,
+} from "../e2e-utils"
 import { caselawTest as test } from "../fixtures"
-import { createDataTransfer } from "~/e2e/shared/e2e-utils"
 
 test.describe("upload an original document to a doc unit", () => {
   test.beforeEach(async ({ page, documentNumber }) => {
